+++ tomcat/src/main/java/org/apache/coyote/http11/common/HttpProtocolVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.http11.common;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
