+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponseStatusLine.java
@@ -2,6 +2,7 @@ package org.apache.coyote.http11.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.apache.coyote.http11.common.HttpProtocolVersion;
 
 @Getter
 @RequiredArgsConstructor
